requirements.txt
youtube-transcript-api
groq
python-dotenv
